Add Shift_N functions, add scrap SQLite3 code
Made functions to move  a state by either up,down,left, or right, by
N points. This function doesn't check for runoff just simplies mods
the shifted value by the axis that its on. There are two ways to handle
runoff, one is place state at edge in the event of runoff, or to make
the function return bool if the state will runoff. Also, there's a little
sqlite3 code that needs more to work. The SELECT statement is what is
needed to store and parse results.