Add StdPieceSet class, clean up code
Added a StdPieceSet class that holds the standard set of pieces for
a game. Right now, the container is an array, but depending on the
rules of the game a dynamic container might be needed. Also, the
structs XY_Indices and XYZ_Indices have been renamed to Point and
Point3 respectively.