Implement more piece moving functionality
I've added an Orientation enum inside the Board class to differentiate
how one access the elements on a Matrix3. If one wants the x,z to in
front of your left shoulder, then use ::Positive, if one wants the
opposite, then use ::Negative. This should moved to Matrix3. The
code is sloppy and needs a more stable handling. Files and classes
need a more clear purpose and shouldn't do more than needed.