Reorganize codebase to be more clean
Reorganized the code base so that the things or functionalities that
are involved in the ruleset or basic structure of the game are stored
in the Protocol.hpp/cpp file. Renamed Gungi.hpp/cpp to Engine.hpp/cpp
as it's a better suited name. This file will handle game and player
states using Protocol.hpp/cpp as its main interface. Also, I removed
MatrixManip.hpp and FixedVector.hpp as I don't feel as I didn't feel
as if they were needed. The Board is no longer a subclass of Matrix3
but rather an alias specialization of:
    Matrix3<const IndexedPiece*, uint8_t>.
The size on the second parameter is a feature added to both matrices
classes so as to save more space.