Allow for pieces to be dropped, during placement
Tested the validations of pieces attempted to be placed. The
pieces can only be placed in the first 3 rows, as per the ruleset.
Also, a soldier must occupy every column during placement phase.
I've implemented a helper function to initialize the pieces
in a predetermined setup for debugging. The pieces do get placed
correctly. The board print function needs work though. Before moving
on to ironing out the engine while the game is running (Phase::Running)
, I must run some more tests on the placement phase. Note, there are
a bunch of print statement scattered around the code base, disregard
them as they are are for debugging.